Checking object existence for a name specified

+++ rdm-database/src/main/procedures/p_delete_meta_type.sql
@@ -18,6 +18,10 @@ begin
 	where
 		t.internal_name = i_internal_name
 	;
+
+	if l_type_id is null then 
+		raise exception 'The type specified is invalid: %', i_internal_name; 
+	end if;
 	
 	execute format('
 		drop table if exists ${stagingSchemaName}.%I
@@ -39,48 +43,46 @@ begin
 		, i_internal_name
 	);
 	
-	if l_type_id is not null then
-		delete from ${mainSchemaName}.meta_index_column 
-		where master_id in (
-			select 
-				i.id
-			from 
-				${mainSchemaName}.meta_index i
-			where
-				i.master_id = l_type_id
-		)
-		;
+	delete from ${mainSchemaName}.meta_index_column 
+	where master_id in (
+		select 
+			i.id
+		from 
+			${mainSchemaName}.meta_index i
+		where
+			i.master_id = l_type_id
+	)
+	;
+
+	delete from ${mainSchemaName}.meta_index
+	where master_id = l_type_id
+	;
 	
-		delete from ${mainSchemaName}.meta_index
-		where master_id = l_type_id
-		;
-		
-		delete from ${mainSchemaName}.meta_attribute_lc 
-		where master_id in (
-			select 
-				a.id
-			from 
-				${mainSchemaName}.meta_attribute a
-			where
-				a.master_id = l_type_id
-		)
-		;
-		
-		delete from ${mainSchemaName}.meta_attribute a
-		where master_id = l_type_id
-		;
+	delete from ${mainSchemaName}.meta_attribute_lc 
+	where master_id in (
+		select 
+			a.id
+		from 
+			${mainSchemaName}.meta_attribute a
+		where
+			a.master_id = l_type_id
+	)
+	;
 	
-		delete from ${stagingSchemaName}.data_package
-		where type_id = l_type_id
-		;
-	
-		delete from ${mainSchemaName}.meta_type_lc 
-		where master_id = l_type_id
-		;
-	
-		delete from ${mainSchemaName}.meta_type 
-		where id = l_type_id
-		;
-	end if;  
+	delete from ${mainSchemaName}.meta_attribute a
+	where master_id = l_type_id
+	;
+
+	delete from ${stagingSchemaName}.data_package
+	where type_id = l_type_id
+	;
+
+	delete from ${mainSchemaName}.meta_type_lc 
+	where master_id = l_type_id
+	;
+
+	delete from ${mainSchemaName}.meta_type 
+	where id = l_type_id
+	;
 end
 $procedure$;
